+++ Nethereum.BlockchainProcessing/Processing/BlockNotFoundException.cs
@@ -1,10 +1,11 @@
 ﻿using System;
+using System.Numerics;
 
 namespace Nethereum.BlockchainProcessing.Processing
 {
     public class BlockNotFoundException: Exception
     {
-        public BlockNotFoundException(ulong blockNumber):
+        public BlockNotFoundException(BigInteger blockNumber):
             base($"Block is null - assumed to not yet exist.  BlockNumber: {blockNumber}"){}
     }
 }
