+++ Storage/Nethereum.BlockchainStore.Csv/Repositories/BlockRepository.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 using CsvHelper.Configuration;
@@ -22,13 +23,13 @@ namespace Nethereum.BlockchainStore.Csv.Repositories
             return await FindAsync(b => b.BlockNumber == blockNumber.Value.ToString());
         }
 
-        public async Task<ulong?> GetMaxBlockNumberAsync()
+        public async Task<BigInteger?> GetMaxBlockNumberAsync()
         {
-            ulong? maxBlock = null;
+            BigInteger? maxBlock = null;
 
             await EnumerateAsync(e =>
             {
-                var blockNumber = ulong.Parse(e.BlockNumber);
+                var blockNumber = BigInteger.Parse(e.BlockNumber);
 
                 if (maxBlock == null || maxBlock.Value < blockNumber)
                 {
